+++ CHANGELOG.md
@@ -9,13 +9,14 @@
 - jslint - simplify comments/docs by removing unnecessary grammar-article "the".
 - jslint - try to improve parser to be able to parse jquery.js without stopping.
 - jslint-refactor - migrate recursive-loops to for/while loops.
+- merge function.html and help.html into README.md
 - node - after node-v12 is deprecated, change `require("fs").promises` to `require("fs/promises")`.
 - node - after node-v14 is deprecated, remove shell-code `export "NODE_OPTIONS=--unhandled-rejections=strict"`.
 - tests - update function warn_at() with assertion-check matching column with artifact.
 
 ## v2021.6.24-beta
+- breaking-change - rename files *.js to *.mjs for better integration with nodejs.
 - ci - include explicit commonjs (jslint.cjs) and es-module (jslint.mjs) variants of jslint.
-- jslint - rename files *.js to *.mjs for better integration with nodejs.
 
 ## v2021.6.22
 - bugfix - fix global_list being ignored by jslint.